requirements.txt
Flask==2.0.1
Werkzeug==2.0.3
Pillow==9.0.0
psycopg2==2.9.3
gunicorn==20.1.0